// src/integrations/supabase/types.ts
export type Database = {
  public: {
    Tables: {
// ...
      users: {
        Row: {
          age: number
          contact: string
          created_at: string
          dni: string
          employment_id: string
          gender: string
          id: string
          is_active: boolean
          name: string
          pathology_id: string
          photo: string | null
          updated_at: string
        }
        Insert: {
          age: number
          contact: string
          created_at?: string
          dni: string
          employment_id: string
          gender: string
          id?: string
          is_active?: boolean
          name: string
          pathology_id: string
          photo?: string | null
          updated_at?: string
        }
        Update: {
          age?: number
          contact?: string
          created_at?: string
          dni?: string
          employment_id?: string
          gender?: string
          id?: string
          is_active?: boolean
          name?: string
          pathology_id?: string
          photo?: string | null
          updated_at?: string
        }
        Relationships: [
          {
            foreignKeyName: "users_employment_id_fkey"
            columns: ["employment_id"]
            isOneToOne: false
            referencedRelation: "employments"
            referencedColumns: ["id"]
          },
          {
            foreignKeyName: "users_pathology_id_fkey"
            columns: ["pathology_id"]
            isOneToOne: false
            referencedRelation: "pathologies"
            referencedColumns: ["id"]
          },
        ]
      }
    }
    Views: {
      [_ in never]: never
    }
    Functions: {
      [_ in never]: never
    }
    Enums: {
      [_ in never]: never
    }
    CompositeTypes: {
      [_ in never]: never
    }
  }
}